+++ db/migrate/20190305084605_create_transaction.rb
@@ -4,15 +4,15 @@ class CreateTransaction < ActiveRecord::Migration[5.2]
     	t.string  :invoice, null: false
     	t.references :user, foreign_key: true, null: false
     	t.references :member, foreign_key: true
-    	t.float :total, null: false
-    	t.float :discount, default: 0
-    	t.float :grand_total, null: false
+    	t.bigint :total, null: false
+    	t.bigint :discount, default: 0
+    	t.bigint :grand_total, null: false
     	t.integer :items, null: false
     	t.integer :payment_type, default: 1
     	t.integer :bank, default: 0
-    	t.float :edc_inv, default: 0
+    	t.bigint :edc_inv, default: 0
         t.timestamp :date_created, null: false
-        t.float :hpp_total, null: false, default: 0
+        t.bigint :hpp_total, null: false, default: 0
 
         t.timestamps
     end
