+++ db/migrate/20190609130722_create_grocier_item.rb
@@ -4,8 +4,8 @@ class CreateGrocierItem < ActiveRecord::Migration[5.2]
     	t.references :item, foreign_key: true, null: false
     	t.integer :min, null: false
     	t.integer :max, null: false
-    	t.float :price, null: false
-        t.float :discount, null: false, default: 0
+    	t.bigint :price, null: false
+        t.bigint :discount, null: false, default: 0
     	
     	t.timestamps
     end
