+++ db/migrate/20190823081123_create_report.rb
@@ -3,15 +3,15 @@ class CreateReport < ActiveRecord::Migration[5.2]
     create_table :reports do |t|
     	t.references :store, foreign_key: true, null: false
 
-  		t.float :cash, null: false, default: 0
-  		t.float :stock_value, null: false, default: 0
-    	t.float :receivable, null: false, default: 0
-    	t.float :asset, null: false, default: 0
+  		t.bigint :cash, null: false, default: 0
+  		t.bigint :stock_value, null: false, default: 0
+    	t.bigint :receivable, null: false, default: 0
+    	t.bigint :asset, null: false, default: 0
 
-		  t.float :capital, null: false, default: 0
-    	t.float :debt, null: false, default: 0
-		  t.float :outcome, null: false, default: 0
-		  t.float :sales, null: false, default: 0
+		  t.bigint :capital, null: false, default: 0
+    	t.bigint :debt, null: false, default: 0
+		  t.bigint :outcome, null: false, default: 0
+		  t.bigint :sales, null: false, default: 0
 
       t.timestamps
